+++ scripts/tiles/utils.ts
@@ -1,4 +1,5 @@
-import { readFile, rename } from 'fs/promises';
+import { readdir, readFile, rename, stat } from 'fs/promises';
+import { join } from 'path';
 
 import geojsonvt from 'geojson-vt';
 import vtpbf from 'vt-pbf';
@@ -18,6 +19,77 @@ const globalX13Min = 3900;
 const globalX13Max = 4400;
 const globalY13Min = 2700;
 const globalY13Max = 3100;
+
+/**
+ * Generate tiles from GeoJSON data and store them in a postgres table
+ */
+export const importMvtDirectory = async (basePath: string, destinationTable: string) => {
+  const startTime = Date.now();
+
+  const zoomLevels = await listDirectoryEntries(basePath, 'dir');
+
+  const tiles: { x: number; y: number; z: number; path: string }[] = [];
+  await Promise.all(
+    zoomLevels.map(async (zoomLevel) => {
+      const columns = await listDirectoryEntries(join(basePath, zoomLevel), 'dir');
+      await Promise.all(
+        columns.map(async (column) => {
+          const rows = await listDirectoryEntries(join(basePath, zoomLevel, column), 'file');
+          rows.forEach((row) => {
+            tiles.push({
+              x: Number.parseInt(column),
+              y: Number.parseInt(row),
+              z: Number.parseInt(zoomLevel),
+              path: join(basePath, zoomLevel, column, row),
+            });
+          });
+        })
+      );
+    })
+  );
+
+  logger.info('importing tiles', {
+    nbTiles: tiles.length,
+  });
+
+  await processInParallel(tiles, QUERY_PARALLELISM, async (tile) => {
+    try {
+      const tileData = await readFile(tile.path);
+      await db(destinationTable)
+        .insert({
+          x: tile.x,
+          y: tile.y,
+          z: tile.z,
+          tile: tileData,
+        })
+        .onConflict(['x', 'y', 'z'])
+        .merge();
+    } catch (err: any) {
+      logger.error(`Error inserting tile ${tile.z}/${tile.x}/${tile.y}`, {
+        error: err.message,
+      });
+    }
+  });
+
+  logger.info('finished importing tiles', {
+    duration: Date.now() - startTime,
+  });
+};
+
+async function listDirectoryEntries(basePath: string, type: 'dir' | 'file'): Promise<string[]> {
+  const entries = await readdir(basePath);
+  const subEntries: string[] = [];
+
+  await Promise.all(
+    entries.map(async (name) => {
+      const fileStats = await stat(join(basePath, name));
+      if ((type === 'dir' && fileStats.isDirectory()) || (type === 'file' && fileStats.isFile())) {
+        subEntries.push(name);
+      }
+    })
+  );
+  return subEntries;
+}
 
 const geoJSONQuery = (properties: string[], id: string) =>
   db.raw(
@@ -269,6 +341,28 @@ export const generateGeoJSON = async (filepath: string) => {
   await rename(`${dockerVolumePath}/output.geojson`, filepath);
 
   return filepath;
+};
+
+export const importMvtDirectoryToTable = async (mvtDirectory: string, destinationTable: string) => {
+  if (await db.schema.hasTable(destinationTable)) {
+    logger.info('flushing destination table', {
+      table: destinationTable,
+    });
+    await db(destinationTable).delete();
+  } else {
+    logger.info('destination table does not exist, creating it', {
+      table: destinationTable,
+    });
+    await db.schema.createTable(destinationTable, (table) => {
+      table.bigInteger('x').notNullable();
+      table.bigInteger('y').notNullable();
+      table.bigInteger('z').notNullable();
+      table.specificType('tile', 'bytea').notNullable();
+      table.primary(['x', 'y', 'z']);
+    });
+  }
+
+  await importMvtDirectory(mvtDirectory, destinationTable);
 };
 
 /**
